# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khadamatak SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /followups:
    get:
      operationId: listFollowups
      tags: [followups]
      summary: List follow-ups
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Follow-ups list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowupsListResponse"
    post:
      operationId: createFollowup
      tags: [followups]
      summary: Create follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowupResponse"
...
components:
  schemas:
...
    Followup:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        contactName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        note:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, status]
    FollowupResponse:
      type: object
      properties:
        followup:
          $ref: "#/components/schemas/Followup"
    FollowupsListResponse:
      type: object
      properties:
        followups:
          type: array
          items:
            $ref: "#/components/schemas/Followup"
        total:
          type: integer
    CreateFollowupBody:
      type: object
      properties:
        title:
          type: string
        contactId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        note:
          type: string
      required: [title, scheduledAt]
...